Adding generation utils
